Move authMiddlewares v2-specific code into a new authMiddlewares only for v2
